feat: SplitViewController 시작 페이지로 설정
