--- CloudNotes/CloudNotes/SplitViewController.swift
@@ -1,29 +1,0 @@
-//
-//  CloudNotes - ViewController.swift
-//  Created by yagom. 
-//  Copyright © yagom. All rights reserved.
-// 
-
-import UIKit
-
-class SplitViewController: UISplitViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        let test = UILabel()
-        view.backgroundColor = .white // 배경색
-        view.addSubview(test)
-        test.text = "text" // test를 위해서 출력할 라벨
-        test.textColor = .black
-        test.translatesAutoresizingMaskIntoConstraints = false
-        test.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-        test.centerYAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
-        
-        self.delegate = self
-        self.preferredDisplayMode = .automatic
-    }
-}
-
-extension SplitViewController: UISplitViewControllerDelegate {
-    
-}
